requirements.txt
pyrogram
TgCrypto
pyromod
python-telegram-bot